+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/SampleAliquotImpl.java
@@ -7,6 +7,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
+import uk.ac.bbsrc.tgac.miso.core.data.BarcodableVisitor;
 import uk.ac.bbsrc.tgac.miso.core.data.SampleAliquot;
 import uk.ac.bbsrc.tgac.miso.core.data.SamplePurpose;
 
@@ -36,4 +37,8 @@ public class SampleAliquotImpl extends DetailedSampleImpl implements SampleAliqu
     return "SampleAliquotImpl [samplePurpose=" + samplePurpose + "]";
   }
 
+  @Override
+  public <T> T visit(BarcodableVisitor<T> visitor) {
+    return visitor.visitSampleAliquot(this);
+  }
 }